rest correction for crc16

+++ EGTS-parcer/src/main/java/ru/tnt/EGTSparser/crc/service/CRC.java
@@ -2,7 +2,7 @@ package ru.tnt.EGTSparser.crc.service;
 
 public interface CRC {
 
-    long calculate8(long[] data);
-    long calculate16(long[] data);
+    long calculate8(byte[] data);
+    long calculate16(byte[] data);
 
 }
